Almost completely Dockerized
One remaining issue is that we get a connection refused
error response to the REST API caller from the gRPC service
when the gRPC service and the REST Proxy are both running
in the Docker container.

There are no issues making a gRPC call to the gRPC
service running in the container from a gRPC client
running outside of the container, and there are no issues
making a REST call to the gRPC service when the gRPC
service and the REST Proxy are both running outside of
the container.